@M-PC 2020-03-16
Shut up just let it be...

On 17 Mar 2020 3:11 a.m., "Mazdak Farrokhzad" <notifications@github.com>
wrote:

> @bors <https://github.com/bors> rollup=never
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/rust-lang/rust/pull/70052#issuecomment-599712425>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AGJEPHCUF2H2LIBNTV7GAZTRHZ2UJANCNFSM4LMRCQMA>
> .
>