@Mark-Simulacrum 2020-03-17
Looking over the diff in the PR you linked, this adds two uses of specialization to hashbrown which I believe carry over into std. I believe we are currently trying to vet specialization before adding more copies by ccing lang folks.

 - https://github.com/rust-lang/hashbrown/pull/146/files#diff-178822c448ce6afe0e383be6f514626eR215-R233
 - https://github.com/rust-lang/hashbrown/pull/146/files#diff-1b47db5c9a30eca8cfa833e56ecf2c4eR1034-R1066

I am unsure who the "experts" here are, but let's try @nikomatsakis and perhaps @Centril first?