@Centril 2020-03-17
Aside: reliance on `default fn` as a syntax here is implemented in a problematic way. The right way to do it is to delay parsing in a macro or move the `default fn`s to a separate external module that is cfged. The way this is done in the diffs above, we will have a headache if we ever decide to change the syntax for specialization, so this should be fixed.