@Amanieu 2020-03-17
Could you give an example of how delayed parsing with a macro would work?

Is it something like this?
```rust
#[cfg(feature = "nightly")]
macro_rules! default_fn {
	($($x:tt)*) => {
        default $($x)*
    }
}
#[cfg(not(feature = "nightly"))]
macro_rules! default_fn {
	($($x:tt)*) => {
        $($x)*
    }
}
```