@Centril 2020-03-17
@Amanieu Example due to @petrochenkov https://github.com/rust-lang/rust/issues/65860#issuecomment-548101030.