@Amanieu 2020-03-17
I started a PR at https://github.com/rust-lang/hashbrown/pull/147, let me know if any other changes are needed.