@matthewjasper 2020-03-17
`min_specialization` is in the next nightly.
I would prefer either `min_specialization` to be used if possible or a sepatate feature flag if not.