@Amanieu 2020-03-20
If I'm understanding the `min_specialization` documentation correctly, this means that we need to add `#[rustc_specialization_trait]` on `Copy`, `Eq`, `Hash` and `BuildHasher`. Is that correct?