@nikomatsakis 2020-04-10
I'm trying to remember under what conditions adding `rustc_specialization_trait` makes sense. I'm a bit nervous about us adding it loosely. But I guess I'll wait until @matthewjasper weighs in -- or else I should go refresh my memory perhaps of what precisely it means and when it's ok to use it. =)