@Amanieu 2020-04-10
The specializations are really the only significant improvement that is relevant to libstd. There's not much point in upgrading otherwise.