@matthewjasper 2020-04-10
`rustc_specialization_trait` can't be added to stable traits since it prevents implementations in crates which don't have `feature(min_specialization)` or `feature(specialization)`.
`Copy` might get `rustc_unsafe_specialization_marker` on it temporarily, but it's not really correct and i don't want crates outside of `std` rely on it when it's removed.