@Amanieu 2020-04-27
Hmm, upon further review I do believe that the specialization on `Eq + Hash + BuildHasher` is unsound. I reverted it in hashbrown 0.7.2.

The other one uses `Copy`, which could be made to work with `rustc_unsafe_specialization_marker`. I think it should be fine to rely on this since:
- this is only used when the "nightly" feature of hashbrown is used. It is acceptable for compiler updates to break people who have opted into using nightly features.
- the use of specialization is an implementation detail and can be reverted at any time without any change to the API.
- hashbrown is maintained by the libs team and can quickly react to any changes in the compiler regarding specialization.