@nikomatsakis 2020-04-28
I don't think we should add the specialization marker to `Copy`. It is definitely possible, for better or worse, for people to write impls like

```rust
impl Copy for Foo<'static> { }
```