@Amanieu 2020-04-28
Specializations on `Copy` vs `Clone` are already used in the standard library: https://github.com/rust-lang/rust/blob/master/src/liballoc/vec.rs#L2168