@matthewjasper 2020-05-17
I would rather not have out-of-tree specializations on `Copy` because it would make it harder to change how specialization on `Copy` is done (e.g. an `is_always_applicable_copy` intrinsic).