@Amanieu 2020-06-13
Adding the specialization for `Copy` can make cloning a `HashMap` 3-4x faster:

```
# Without specialization
test clone_from_large            ... bench:       1,506 ns/iter (+/- 109)
test clone_from_small            ... bench:          18 ns/iter (+/- 2)
test clone_large                 ... bench:       2,036 ns/iter (+/- 242)
test clone_small                 ... bench:          46 ns/iter (+/- 7)

# With specialization
test clone_from_large            ... bench:         441 ns/iter (+/- 54)
test clone_from_small            ... bench:           9 ns/iter (+/- 1)
test clone_large                 ... bench:         480 ns/iter (+/- 67)
test clone_small                 ... bench:          30 ns/iter (+/- 2)
```

I think this improvement is significant enough to justify the use of specialization. Also note that hashbrown is maintained by the libs team so its use of specialization won't block any changes to how the compiler does specialization.