@matthewjasper 2020-06-20
I don't think that I'll get around to adding a new way of specializing on `Copy` any time soon. Adding only the `Copy` specialization seems OK.