@Amanieu 2020-06-21
I have updated hashbrown to 0.8.0 which uses `min_specialization` and has a few other optimizations.