@Amanieu 2020-06-21
@ortem: Recent changes to `hashbrown` have broken the debug pretty printers. The `data` field has been removed and key/value pairs are now indexed at a negative offset from `ctrl`.

```
KV3, KV2, KV1, KV0, C0, C1, C2, C3
                    ^ ctrl points here
```

I'm not too familiar with the debug visualizers so I would appreciate a bit of help getting them fixed!